Refactor project structure and fix module loading
- Flattened the project structure by moving all files from the `prueba.turno` subdirectory to the root.
- Merged dependencies from multiple `package.json` files into one.
- Corrected all relative paths in HTML and JavaScript files to work with the new structure.
- Fixed JavaScript modules by ensuring `config.js` exports its configuration object and that all scripts are loaded with `type="module"` where necessary.
- Removed redundant Supabase client initializations.
