+++ admin/panel.js
@@ -1,6 +1,6 @@
 // panel.js
-import { supabase } from '../../../database.js';
-import Config from '../../../config.js';
+import { supabase } from '../database.js';
+import Config from '../config.js';
 
 let negocioId; // Se obtendrá del usuario autenticado
