+++ cierre.html
@@ -8,13 +8,13 @@
 
 
  <!-- Base estático -->
- <base href="" />
+ <base href="./" />
 
-  <link rel="icon" type="image/x-icon" href="../imegenlogin/favicon.ico" />
-  <link rel="icon" type="image/png" href="../imegenlogin/favicon-16x16.png" sizes="16x16" />
-  <link rel="icon" type="image/x-icon" href="./imegenlogin/favicon-32x32.png" sizes="32x32" />
-  <link rel="apple-touch-icon" href="./imegenlogin/apple-touch-icon.png" />
-  <link rel="shortcut icon" href="./imegenlogin/favicon.ico" type="image/x-icon"/>
+  <link rel="icon" type="image/x-icon" href="imegenlogin/favicon.ico" />
+  <link rel="icon" type="image/png" href="imegenlogin/favicon-16x16.png" sizes="16x16" />
+  <link rel="icon" type="image/png" href="imegenlogin/favicon-32x32.png" sizes="32x32" />
+  <link rel="apple-touch-icon" href="imegenlogin/apple-touch-icon.png" />
+  <link rel="shortcut icon" href="imegenlogin/favicon.ico" type="image/x-icon"/>
