+++ config.js
@@ -64,3 +64,5 @@ const Config = {
     };
   }
 };
+
+export default Config;
