+++ create_negocio_config_table.sql
