+++ package.json
@@ -9,7 +9,8 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "lit": "^3.3.0"
+    "lit": "^3.3.0",
+    "@supabase/supabase-js": "^2.53.0"
   },
   "devDependencies": {
     "vite": "^7.0.4"
