--- prueba.turno/package.json
@@ -1,12 +1,0 @@
-{
-  "dependencies": {
-    "@supabase/supabase-js": "^2.53.0"
-  },
-  "devDependencies": {
-    "javascript-obfuscator": "^4.1.1",
-    "terser": "^5.43.1"
-  },
-  "scripts": {
-    "build": "node build.js"
-  }
-}
