+++ registraringresos.js
