+++ turno.html
@@ -6,13 +6,13 @@
   <title>Gestión de Turnos - TurnoRD</title>
 
   <!-- Base estático -->
-  <base href="" />
+  <base href="./" />
 
-  <link rel="icon" type="image/x-icon" href="../imegenlogin/favicon.ico" />
-  <link rel="icon" type="image/png" sizes="16x16" href="../imegenlogin/favicon-16x16.png" />
-  <link rel="icon" type="image/png" sizes="32x32" href="../imegenlogin/favicon-32x32.png" />
-  <link rel="apple-touch-icon" sizes="180x180" href="../imegenlogin/apple-touch-icon.png">
-  <link rel="manifest" href="../imegenlogin/site.webmanifest">
+  <link rel="icon" type="image/x-icon" href="imegenlogin/favicon.ico" />
+  <link rel="icon" type="image/png" sizes="16x16" href="imegenlogin/favicon-16x16.png" />
+  <link rel="icon" type="image/png" sizes="32x32" href="imegenlogin/favicon-32x32.png" />
+  <link rel="apple-touch-icon" sizes="180x180" href="imegenlogin/apple-touch-icon.png">
+  <link rel="manifest" href="imegenlogin/site.webmanifest">
