+++ usuario.html
@@ -6,13 +6,13 @@
 <title>TurnORD - Usuario</title>
 
 <!-- Base estático -->
-<base href="" />
+<base href="./" />
 
 <!-- Favicons -->
-<link rel="icon" type="image/x-icon" href="../imegenlogin/favicon.ico" />
-<link rel="icon" type="image/png" sizes="16x16" href="../imegenlogin/favicon-16x16.png" />
-<link rel="icon" type="image/png" sizes="32x32" href="../imegenlogin/favicon-32x32.png" />
-<link rel="apple-touch-icon" sizes="180x180" href="../imegenlogin/apple-touch-icon.png">
+<link rel="icon" type="image/x-icon" href="imegenlogin/favicon.ico" />
+<link rel="icon" type="image/png" sizes="16x16" href="imegenlogin/favicon-16x16.png" />
+<link rel="icon" type="image/png" sizes="32x32" href="imegenlogin/favicon-32x32.png" />
+<link rel="apple-touch-icon" sizes="180x180" href="imegenlogin/apple-touch-icon.png">
 
 <!-- Tailwind CSS -->
 <script src="https://cdn.tailwindcss.com"></script>
@@ -98,9 +98,6 @@
 
   <div id="listaEspera" class="space-y-2 mt-4"></div>
 </div>
-
-<!-- Supabase -->
-<script src="https://cdn.jsdelivr.net/npm/@supabase/supabase-js@2"></script>
 
 <!-- Errores -->
 <script>
